// lib/definitions/product-types.ts
export type Product = {
  id: string;
  name: string;
  description?: string;
  imageSrc: StaticImageData;
  alt: string;
  category?: string;
  sub_category?: string;
  price?: number;
};
// update types to be GlassType = Product & {shape_types: Shape[]}
export type GlassType = Product & {
  shape_types?: Shape[];
  thickness_options?: Thickness[];
  tint_types?: Tint[];
  misc_options?: MiscOptions[];
};

export type Shape = Product & {
  required_dimensions?: Dimension[];
  thickness_options?: Thickness[];
  tint_types?: Tint[];
  misc_options?: MiscOptions[];
};

export type Tint = Product & {
  misc_options?: MiscOptions[];
};

export type Dimension = {
  label: string;
  wholeNumber: string;
  fraction: string;
};

export type Thickness = {
  id?: string;
  thickness?: number;
  tint_types?: Tint[];
  misc_options?: MiscOptions[];
};

export type FabricationOptions = {
  id?: string;
  add_holes?: boolean;
  add_cutouts?: boolean;
  add_notches?: boolean;
  add_mitered_edges?: boolean;
  add_flat_polish?: boolean;
  add_pencil_polish?: boolean;
  add_bevel_edge?: boolean;
  add_tempering?: boolean;
  add_heat_strengthening?: boolean;
  add_laminating?: boolean;
};

export type MiscOptions = {
  id?: string;
  add_tempered_logo?: boolean;
};